+++ src/api/rains/RaainApiRainsCumulativesCreateRequest.ts
@@ -1,0 +1,9 @@
+// POST /rains/:rainId/cumulatives request body
+export interface RaainApiRainsCumulativesCreateRequest {
+    periodBegin: string;
+    periodEnd: string;
+    provider: string;
+    confName?: string;
+    timeStepInMinutes?: number;
+    groundJobIds?: string[];
+}