+++ src/api/rains/RaainApiRainsCumulativesCreateResponse.ts
@@ -1,0 +1,7 @@
+// POST /rains/:rainId/cumulatives response
+export interface RaainApiRainsCumulativesCreateResponse {
+    postJobId: string;
+    getJobId: string;
+    groundJobIds?: string[];
+    cumulativeWindowInMinutes: number;
+}