+++ src/polar/PolarMeasureValue.ts
@@ -2,7 +2,7 @@ import {MeasureValuePolarContainer} from './MeasureValuePolarContainer';
 import {IPolarMeasureValue} from './IPolarMeasureValue';
 import {PolarValue} from './PolarValue';
 import {AbstractPolarMeasureValue} from './AbstractPolarMeasureValue';
-import {calculateMinMax} from '../utils';
+import {Tools} from '../utils';
 
 export class PolarMeasureValue implements IPolarMeasureValue {
     protected measureValuePolarContainers: MeasureValuePolarContainer[];
@@ -345,7 +345,7 @@ export class PolarMeasureValue implements IPolarMeasureValue {
             );
         }
 
-        return calculateMinMax(allValues);
+        return Tools.calculateMinMax(allValues);
     }
 
     public getAzimuthsInDegrees() {